update update-node functions

+++ scripts/update/update-node.sh
@@ -28,9 +28,6 @@ for param in "$@"; do
     esac
 done
 
-
-SKIP_VERSION=$(yq '.scheduled_tasks.updates.node.skip_version // "false"' $QTOOLS_CONFIG_FILE)
-
 CURRENT_NODE_VERSION=$(get_current_node_version)
 CURRENT_QCLIENT_VERSION=$(get_current_qclient_version)
 RELEASE_NODE_VERSION=$(fetch_node_release_version)
@@ -42,6 +39,7 @@ if [ "$auto_update" == "true" ] && [ "$is_auto_update_enabled" == "false" ]; the
   exit 0
 fi
 
+SKIP_VERSION=$(yq '.scheduled_tasks.updates.node.skip_version' $QTOOLS_CONFIG_FILE)
 # otherwise we may want to skip updating to the current version for now, but on the next update we will update it
 if [ "$SKIP_VERSION" != "false" ] && [ "$SKIP_VERSION" != "" ] && [ "$RELEASE_VERSION" == "$SKIP_VERSION" ]; then
   log "Skipping update for version $SKIP_VERSION"
@@ -61,7 +59,7 @@ clean_old_node_files() {
     # Get a list of all files in $QUIL_NODE_PATH that don't match $release_version and remove them
     log "Removing old node files..."
     for file in "$QUIL_NODE_PATH"/*; do
-      if [[ -f "$file" && ! "$file" =~ .*$RELEASE_NODE_VERSION.* ]]; then
+      if [[ -f "$file" && ! "$file" =~ .*$RELEASE_NODE_VERSION.* && ! "$file" =~ .*\.dgst(\.sig\.[0-9]+)?$ ]]; then
         log "Removing old file: $file"
         rm -f "$file"
       fi
@@ -70,7 +68,7 @@ clean_old_node_files() {
     # Get a list of all files in $QUIL_CLIENT_PATH that don't match $release_version and remove them
     log "Removing old qclient files..."
     for file in "$QUIL_CLIENT_PATH"/*; do
-      if [[ -f "$file" && ! "$file" =~ $RELEASE_QCLIENT_VERSION ]]; then
+      if [[ -f "$file" && ! "$file" =~ $RELEASE_QCLIENT_VERSION && ! "$file" =~ .*\.dgst(\.sig\.[0-9]+)?$ ]]; then
         log "Removing old file: $file"
         rm -f "$file"
       fi
